refactor(polar): move sausage to util/shape
