+++ src/util/symbol.js
@@ -130,79 +130,6 @@ var Pin = graphic.extendShape({
 });
 
 /**
- * Sausage: similar to sector, but have half circle on both sides
- * @inner
- */
-export var Sausage = graphic.extendShape({
-
-    type: 'sausage',
-
-    shape: {
-
-        cx: 0,
-
-        cy: 0,
-
-        r0: 0,
-
-        r: 0,
-
-        startAngle: 0,
-
-        endAngle: Math.PI * 2,
-
-        clockwise: true
-    },
-
-    buildPath: function (ctx, shape) {
-        var x = shape.cx;
-        var y = shape.cy;
-        var r0 = Math.max(shape.r0 || 0, 0);
-        var r = Math.max(shape.r, 0);
-        var dr = (r - r0) * 0.5;
-        var rCenter = r0 + dr;
-        var startAngle = shape.startAngle;
-        var endAngle = shape.endAngle;
-        var clockwise = shape.clockwise;
-
-        var unitStartX = Math.cos(startAngle);
-        var unitStartY = Math.sin(startAngle);
-        var unitEndX = Math.cos(endAngle);
-        var unitEndY = Math.sin(endAngle);
-
-        var lessThanCircle = clockwise
-            ? endAngle - startAngle < Math.PI * 2
-            : startAngle - endAngle < Math.PI * 2;
-
-        if (lessThanCircle) {
-            ctx.moveTo(unitStartX * r0 + x, unitStartY * r0 + y);
-
-            ctx.arc(
-                unitStartX * rCenter + x, unitStartY * rCenter + y, dr,
-                -Math.PI + startAngle, startAngle, !clockwise
-            );
-        }
-
-        ctx.arc(x, y, r, startAngle, endAngle, !clockwise);
-
-        ctx.moveTo(unitEndX * r + x, unitEndY * r + y);
-
-        ctx.arc(
-            unitEndX * rCenter + x, unitEndY * rCenter + y, dr,
-            endAngle - Math.PI * 2, endAngle - Math.PI, !clockwise
-        );
-
-        if (r0 !== 0) {
-            ctx.arc(x, y, r0, endAngle, startAngle, clockwise);
-
-            ctx.moveTo(unitStartX * r0 + x, unitEndY * r0 + y);
-        }
-
-        ctx.closePath();
-    }
-});
-
-/**
  * Arrow shape
  * @inner
  */
@@ -252,8 +179,6 @@ var symbolCtors = {
 
     pin: Pin,
 
-    sausage: Sausage,
-
     arrow: Arrow,
 
     triangle: Triangle
@@ -282,15 +207,6 @@ var symbolShapeMakers = {
         shape.width = w;
         shape.height = h;
         shape.r = Math.min(w, h) / 4;
-    },
-
-    sausage: function (x, y, w, h, shape) {
-        shape.r = Math.min(w, h) / 2;
-        shape.r0 = shape.r / 2;
-        shape.cx = x + w / 2;
-        shape.cy = y + h / 2;
-        shape.startAngle = -Math.PI / 6;
-        shape.endAngle = Math.PI / 6 * 7;
     },
 
     square: function (x, y, w, h, shape) {
